+++ generators/app/templates/api/src/routes/index.js
@@ -1,10 +1,10 @@
-import Express from 'express'
+import express from 'express'
 import { tokenRouter } from './authentication/tokenRouter.js'
 import { authenticationRouter } from './authentication/authenticationRouter.js'
 import { registerRouter } from './user/registerRouter.js'
 import { userRouter } from './user/router.js'
 
-const router = Express.Router()
+const router = express.Router()
 
 router.use('/users/register', registerRouter)
 router.use('/tokens', tokenRouter)